Add files via upload
This video uses opencv2 in python for image processing and is for anyone with a knack for image processing.

Hello there, Let's begin with yet another interesting and great project that will add values and demonstrate your skills. But before we start, it would be good if you know basic image processing concepts, however I'll give proper references from time to time, so its all good.